Adds sass_functions attribute which will be passed through to CSS::Sass
CSS::Sass can take a sass_functions option which enables the user to
write perl functions that can be called from inside their sass. This
patch allows the user to set sass_functions in the say way that they can
set include_paths. This is only useful when the CSS::Sass library is
used to compile to css.